First steps toward reworking the simple loop fixpoint feature.
This formulation is based on a more traditional style of loop invariant.
This avoids the need for the execution feature to discover and
characterize the loop indexing structure, as we can rely on the user to
instead do that for us.

In addition, we now can capture writes to SMT array backed regions
of memory so that they can be involved in the loop invariant.