`crucible-llvm`: Match names up to Darwin aliases
On the Darwin family of operating systems, which includes macOS, various
`libc` functions are defined in terms of aliases that contain various odd
prefixes and suffixes. For instance, the `write()` function is actually defined
in terms of a lower-level `\01_write()` function, and if you compile a C program
containing an occurrence of `write()` with Clang on Darwin, the resulting LLVM
bitcode will mention `\01_write()`, not `write()`. This poses an issue for
`crucible-llvm`'s override for `write()`, since it now mentions the wrong name
on Darwin. Moreover, it isn't feasible to make the LLVM bitcode mention
`write()` instead.

To compensate for this oddity, we introduce a new `DarwinAliasMatch` template
matcher and use it in `basic_llvm_override` when compiling on Darwin. This
matcher indicates that a function name is allowed to match the override name
after possibly removing some Darwin-related prefixes or suffixes. See
`Note [Darwin aliases]` for the full story.

Fixes #885.