CI: Actually test crux-llvm on macOS
Previously, the `clang-withIncl` script that we concocted on CI for calling
`clang` with the correct flags neglected to actually pass the _arguments_ to
`clang`. As a result, the `crux-llvm` test suite would trivially pass on macOS
since it skipped all of the test cases due to `clang-withIncl` being
misconfigured. This corrects this mistake by passing the arguments through to
`clang-withIncl` using `${1+"$@"}`.

Fixes #999. This won't be enough on its own to make the `crux-llvm` test suite
pass on macOS—see #885 and #1013. Fixes for those issues will come in
subsequent commits.