crucible-syntax: Add support for externs
`extern` is to `defglobal` what `declare` is to `defun`. This patch adds the
necessary parts to `crucible-syntax` to parse `extern`s, create fresh
`GlobalVar`s for each `extern`, and put them into the `ParsedProgram` API so
that callers can insert externs into the `SymGlobalState` with appropriate
`RegValue`s. I've included a test case that demonstrates this in action.

Fixes #1054.