crucible-llvm: Do not assume UTF-8 encoding in printf override (#1062)
Previously, the implementation of the `printf` override would UTF-8–encode all
string arguments that it displays. This patch changes it so that it instead
converts the characters from each string argument directly to `ByteString`s
using `Data.ByteString.pack`, without picking a particular text encoding. This
relies on the assumption that C strings are arrays of 1-byte characters, which
is an assumption that the implementation was already relying on implicitly. I
have explicitly documented this assumption in
`crucible-llvm/doc/limitations.md`.

It is difficult to write a `crux-llvm` test case that properly covers this new
code, as the `printf` override's output is somewhat insensitive to the exact
text encoding used. This is primarily useful for downstream code that
implements a `sprintf` override, and in that context, the text encoding very
much matters, as it can affect the length of the string that it stores.

Fixes #1010.