Remove uses of NoMonoLocalBinds after GADTs/TypeFamilies pragmas
These produce `-Wgadt-mono-local-binds` warnings with GHC 9.4, which has added
the warning to `-Wall`. To avoid the warning, I have removed any uses of
`NoMonoLocalBinds` in places where they appear in tandem with `GADTs` and/or
`TypeFamilies`.  In some places, `NoMonoLocalBinds` can be removed with little
effort, but in other places, this requires adding some explicit type
signatures.