crucible-wasm: Adapt to multi-value Wasm proposal
This bumps the `haskell-wasm` submodule to include the changes from
https://github.com/SPY/haskell-wasm/commit/b1874dfcea71909bc9e2f36c1ec1caf416a2ebd2
and
https://github.com/SPY/haskell-wasm/commit/b816e04b6df45ae06f5d6990f2a385c31f9e660d,
which change the `Block`, `Loop`, and `If` instructions to use a `BlockType`
instead of a `ResultType`. This allows these instructions to have multiple
result values, as described in this WebAssembly proposal:
https://github.com/WebAssembly/spec/pull/1145

To adapt the code on the `crucible-wasm` side, I have introduced a
`getBlockResultType` function to convert from a `BlockType` to a `ResultType`,
which is heavily inspired by a similar function in `haskell-wasm`. I have also
factored out some code shared in common between `translateFunction` and
`genInstruction`'s case for `Block` so that I do not have to conjure up a fresh
`BlockType` in `translateFunction`.

Part of #1069.