crucible-wasm: Don't make *.config files mandatory
`tasty-sugar`'s `associatedNames` are meant to indicate optional files that can
live alongside the root file, but they shouldn't be mandatory. Indeed, the only
reason that the associated name `config` was mandatory was because of a strange
quirk in the way `crucible-wasm`'s test suite was set up, which I suspect was
an accidental copy-paste leftover from `crux-llvm`'s test suite. Let's remove
this quirk.