Track crate names and their hashes, use it to look up wired-in names
Unlike in the previous Rust nightly, where the disambiguators for crates in the
standard library were always hard-coded to a known constant, the disambiguators
in the new Rust nightly are always unique, even for standard library crates.
As a result, in order to determine what disambiguator a crate like `core` has,
we must look at the MIR JSON and map each crate name to its disambiguator. See
the new `crateHashesMap` field in `CollectionState`.