Mir.Intrinsics: Provide more IO wrappers around memory model functionality
Much of the memory model–related functionality in `Mir.Intrinsics` is defined
in terms of `FancyMuxTree`s, which are somewhat clunky to manipulate outside of
`crucible-mir`. `Mir.Intrinsics` does provide some wrapper functions to
encapsulate the `FancyMuxTree` manipulation, but these typically live in the
`OverrideSim` monad, which is more specific than we want for SAW's needs. This
patch adds `IO` wrappers alongside the `OverrideSim` wrappers to make it more
straightforward to use them in SAW.