Implement overrides for llvm.is.fpclass and friends
This is primarily motivated by Clang 17 being more eager to compile the `isnan`
function into `llvm.is.fpclass`, so this patch is required to make the
`isnan.c` test case pass with Clang 17 or later. The machinery used to
implement an override for `llvm.is.fpclass` can also be used to implement
overrides for the related `isinf` function, so we also do this.

Checks off some boxes in #187.