crux-llvm: Make isinf.c test more robust to different implementations
On Ubuntu, `isinf(x)` will return `1` if `x` is positive infinity and `-1` if
`x` is negative infinity. The C standard doesn't guarantee this, however, as it
only requires that `isinf(x)` return _some_ non-zero value if `x` is infinite.
More to the point, Apple Clang's version of `isinf` doesn't return the same
values that Ubuntu's Clang does, but the `isinf.c` test case in the `crux-llvm`
test suite specifically checks for these values, making the test fail on macOS.

This patch makes the checks in `isinf.c` more general so that they will pass
regardless of which non-zero value gets returned from `isinf(x)`.