CI: Run crux-llvm test suite in macOS
Previous, we concocted a `clang-withIncl` script for running Clang on macOS CI
runners, but this neglected to pass any arguments to Clang (among other
issues). This patch replaces the `clang-withIncl` machinery with a much
simpler, Homebrew-based setup for install Clang/LLVM. With this in place, we
can finally run the `crux-llvm` test suite on macOS.

Fixes #999.