crucible-llvm: Generalize `LLVMOverride`'s `ext` parameter
Previously, the `OverrideSim` action for an `LLVMOverride` was
assumed to have a fixed `ext` parameter, namely, `LLVM`. While this is
intuitively appealing, it turns out not to be necessary. This is because
there are only a few operations in the `OverrideSim` monad that make
essential use of the `ext` parameter, such as registering or calling a
CFG in the given language extension. These operations are unlikely to be
used when defining, e.g., overrides for libc functions.

The motivation for generalizing the parameter is to enable the reuse of
`LLVMOverride`s in other Crucible language extensions that make use
of the LLVM memory model, namely, Macaw.