CI: Run `cabal sdist` as a sanity check
It is possible to misconfigure a package's `.cabal` file in a way that is not
easily detectable unless you try to bundle the package into a source
distribution (e.g., through the `cabal sdist` or `cabal install` commands).
These sorts of misconfigurations have snuck into the `crucible`-related
packages in the past (see #1209 for one such example).

To prevent these sorts of issues from happening in the future, we now run
`cabal sdist` on each package in the CI as a sanity check.