openEventFrames: Replace tail with pattern-matching + panic
We have managed to call `tail` here on empty lists before (see #1000), which
leads to confusing error messages. Let's replace this with explicit
pattern-matching plus a call to `panic` in the empty list case to make it more
obvious where the error comes from. Moreover, doing this fixes a `-Wx-partial`
warning on GHC 9.8, which is also nice.