crucible-wasm: Adapt to select having an optional value type
This bumps the `haskell-wasm` commit to
https://github.com/SPY/haskell-wasm/commit/c743e11ffd621c871ab98068fee4d08f21f42d14
and does the minimum amount of work needed to make `crucible-wasm` continue to
compile after these changes. In particular, the `select` instruction now has an
optional value type indicating the type of its first and second operand, which
`crucible-wasm` ignores. This is OK for now, as `crucible-wasm`'s
implementation is currently simple enough where the types of all its stack
values can be inferred without any additional type information.