crucible-wasm: Adapt to new data segment format in Wasm 2.0
This bumps the `haskell-wasm` commit to
https://github.com/SPY/haskell-wasm/commit/32392dcc29ef3fef3a0e363d3c04d44f3082c881
and does the minimum amount of work needed to make `crucible-wasm` continue
to compile after these changes. For now, we do not implement full support
for Wasm 2.0's new data segment features, as that would require a
non-trivial amount of work to support. See #1228.