llvm: Fix validity predicate for memory reads with symbolic block numbers (#1238)
Previously, this case returned the concretely-true predicate. This is
incorrect, as `isAllocatedGeneric` takes an argument `inAlloc`, which is
supposed to be applied to the `AllocInfo` of the allocation with the
matching block number, as it is in the other branch.

Mea culpa, looks like I introduced this bug long ago.

re: UC-Crux tests. It looks to me like what's happening is that UC-Crux
previously found certain errors in each of these test cases that were
adequately handled by its heuristics. Fixing this predicate introduces a
new kind of error, and the heuristics aren't sophisticated enough to
deal with this one. UC-Crux then gives up, saying it's not sure how to
avoid these errors with a new precondition (leaving them "unclassified").

This is unfortunate, but UC-Crux already can't abduce preconditions for
a wide variety of programs, so I don't consider this a blocker.