Rearrange the way constant references are handled in crucible-mir.
Instead of sending them as blobs with values, mir-json now encodes
them as references to static definitions. The values appear as
separate bodies as static allocations.

The new input kinds "slicebody" and "strbody" are used to ship the
bodies in the JSON.

Change our internal representation of the input MIR to split the
ConstSlice and ConstStr cases accordingly. Kill off the ConstByteStr
case as it's no longer needed; byte strings come through as ordinary
references now.

Adjust the translation step to handle the bodies and references
separately. The critical part of this is that it produces globalMirRef
instead of constMirRef; this avoids using constMirRef for cases it
can't support that lead to "Cannot compare Const_RefRoots". See
saw-script #2064.

Note that the handling of slice and string slice bodies is probably
now the same as the handling for ordinary arrays and could be
simplified away. It is also possible that Const_RefRoots can be killed
off entirely. One thing at a time though.

Requires corresponding update to mir-json.