CI: Make LLVM version selection more fine-grained
ubuntu-22.04 includes LLVM 14, but ubuntu-20.04 does not. As such, we cannot
always install LLVM 14, but rather, we must choose which LLVM version to
install based on which Ubuntu version we are using.

While I am in town, this patch also factors out the LLVM version into a single
`LINUX_LLVM_VER` environment variable and uses it everywhere so that we do not
have to remember to update quite so many locations in the future. (I also do
something similar with `MACOS_LLVM_VER`.)