Dockerfiles: Properly set up non-root users (#1265)
Previously, the `crux-{llvm,mir}` Docker images were set up to use `root`
accounts, and all files that were created in the Docker images were given
`root` permissions. This is impractical for running rootless container
environments, as noted in #1261.

This rewrites the images to set the default users to unprivileges
`crux-{llvm,mir}` users. This is surprisingly involved, and it requires a fair
bit of refactoring to accomplish:

1. We cannot easily base the `crux-mir` image on the official `rust` image,
   as the latter has `cargo`, `rustup`, and related tools installed under a
   `root` account. Instead, we change to an unprivileged user and install these
   Rust tools using rustup.sh's installer script.
2. Previously, the `crux-llvm` image was supporting Unicode character printing
   by changing the locale, which requires root access to accomplish. This is
   overkill, however, as it suffices to simply set the `LANG` and `LC_ALL`
   environment variables. I've opted for the latter approach, which is much
   simpler and avoids the need for root access.
3. I needed to reorganize the order of commands in the `base` layers of each
   image in order to ensure that the relevant files are `chown`ed from the
   unprivileged user accounts.

Fixes #1261.