crucible-mir: Improve pretty-printing of ZST constants
`ConstZST` (i.e., constant values with zero-sized types) do not include any
information about what the zero-sized type is. In particular, function
constants are represented as `ConstZST`s, and the only way to determine what
function the `ConstZST` corresponds to is to look at its type (`FnDef`), which
stores the `DefId` of the function name itself.

Previously, the `crucible-mir` pretty-printer was printing `ConstZST`s as
`<ZST>` because it did not have access to any type information when
pretty-printing `ConstVal`s. This results in very suboptimal pretty-printed
output (see the example below). This is also true for other forms of
`ConstZST`s, although currently `mir-json` only turns function constants into
`ConstZST`s. (In the future, this could conceivably change.)

To fix this issue, we add special cases in the `Pretty Constant` instance
(which does have access to type information for `ConstVal`s) such that
`ConstZST`s are printed using their type information. In particular, if the
`ConstZST` has type `TyFnDef`, then the function `DefId` stored in the
`TyFnDef` is printed. All other forms of `ConstZST`s are printed like `<ZST:
[ty]>`, where `[ty]` is the pretty-printed type.

As an example of the improvements in this patch, consider this program:

```rs
pub fn foo() {}
pub fn bar() { foo() }
```

Before the changes in this patch, the MIR for `bar()` would be pretty-printed
as:

```rs
fn test/f3b25c00::bar[0]() -> () {
   let mut _0 : ();
   bb0: {
      call(_0 = <ZST>(), bb1)
   }
   bb1: {
      return;
   }
}
```

(Note the call to `<ZST>()`, which gives no indication whatsoever about what
function was called.)

After the changes in this patch, it is pretty-printed as:

```rs
fn test/4717ccf0::bar[0]() -> () {
   let mut _0 : ();
   bb0: {
      call(_0 = test/4717ccf0::foo[0](), bb1)
   }
   bb1: {
      return;
   }
}
```

Fixes GaloisInc/mir-json#79.