Add an EqF TypeRepr instance
There are many more `EqF` instances that we _could_ define in `crucible`, but
defining them all would likely require coming up with a solution for
GaloisInc/parameterized-utils#175 first. Until then, this patch adds an `EqF`
instance for `TypeRepr` in particular, which is straightforward to define
without any additional functionality and addresses the needs of a downstream
project.