Bump everything to devel versions and add "next" sections to changelogs.
Do this for everything, not just the subpackages that had devel
versions in the past. Many of the ones we don't issue formal releases
for were at 0.1. For these it's far from clear whether the next
release should theoretically be 0.1 because they haven't been released
yet, or 0.2 because what's there now is more or less 0.1. Probably
there are some of each. Now the next release for all of them will be
0.2, and if this means some never have a version 0.1, that doesn't
hurt anything...