ci: Build packages, not just libraries or executables
Previously, we individually specified various libraries and executables
(via `lib:` and `exe:` prefixes). This is unnecessary, as building the
package should build both the library and any executables. Furthermore,
I think this caused us to not build the dependencies of test suites
when building with `--only-dependencies`. So, just build whole packages
instead.