ci: Revert "Build modules of local packages in parallel"
This reverts commit 4b832c9626906ea40d96ed60906bbe993182195f.

It appears that building with parallelism is more complex than I
anticipated in a multi-package repo. I tried to split the build of
dependencies (which should have their packages built in parallel but
modules built in serial, which is the current behavior for all packages)
from that of local packages (which should be built serially but with
their modules compiled in parallel, which latter part would be new as
of this PR), but that's hard when you have local packages depending on
one another. For instance, `cabal build --only-dependencies exe:crux-go`
builds `crucible-go` (which is local). Thus, I will instead content
myself with building local packages in parallel until we can use
`--semaphore`.