crucible-mir: symlink mir-json/rust-toolchain.toml into crux-mir/
The crux-mir test suite compiles some tests with normal rustc so it can
check that running the test under Crucible produces the same results as
running it natively.  If the default `rustc` is a stable release, the
native build will fail on tests that use the `#![feature]` attribute.
Using the mir-json `rust-toolchain.toml` ensures that we get a nightly
`rustc`, which will support `#![feature]`.