crucible-mir: fix "standalone use of `dyn`" error
Previously, a pointer-to-pointer `Misc` cast would sometimes call
`tyToRepr` on the source and destination pointee type.  If the pointee
type was `TyDynamic` (`dyn Trait`), then `tyToRepr` would throw an
error, since `dyn` normally should only appear behind a pointer or
reference.  This commit adjusts the `Misc` cast to call `tyToRepr` on
the pointer types instead of the pointees.