crucible-mir: Remove MirPlaceDynRef
Now that we can store a trait object's data directly in a `MirPlace`, there is
no longer any need for the `MirPlaceDynRef` hack that we used before. This
removes `MirPlaceDynRef` and replaces it with the new `DynMeta` constructor of
`PtrMetadata`, which is used to store the trait object's vtable.

Fixes #1092.