crucible-mir: Support more UnsizeVtable casts
This extends `evalCast'` to support the following unsized cast types:

1. `UnsizeVtable` casts between two structs. This is needed to cast between,
   say, `Foo<i32>` and `Foo<dyn Bar>`.
2. `UnsizeVtable` casts between two raw pointer types.

This also bumps the `mir-json` submodule to bring in the changes from
GaloisInc/mir-json#122. This is needed to ensure that `mir-json` accurately
classifies `UnsizeVtable` casts between structs, which is necessary to support
the first type of cast listed above.

Fixes #1433.