MIR: recoverably fail on unsupported type (#1461)
* MIR: recoverably fail on unsupported type

The updates tyToRepr to report a recoverable error when we encounter a type we don't support.

This allows us to emit a symbolic-execution runtime assertion failure for code that uses types we don't support. Treating these as soft failures allow us to verify other aspects of the program that don't execute the unsupported portion both through assumption and through overrides.