crucible-mir: Translate "constant" statics before non-constant ones
`mir-json` emits two types of static item definitions: ones with constant
initializer expressions, and ones without them. It is possible to write a
program where a non-constant static item's definition depends on a constant
static item, and in order for `crucible-mir` to handle such a program, it is
important that `crucible-mir` translate the constant static item _before_ the
non-constant static item that uses it.

This is difficult to arrange for in the general case (see #1108), but in the
common case, we can make this much more likely to work by always translating
every constant static item before every non-constant static item. This provides
a partial workaround (but not a full fix) for #1108.