crucible-mir: `transTerminatorKind` discriminates on callees by type
Previously, `transTerminatorKind` would accept `Call`s of `TyFnDef` values only
of a particular (`Constant`) shape, but it turns out `rustc` can generate MIR
that refers to them as e.g. variables. (I am not sure whether this is a change
in `rustc`'s behavior, or if this shortcoming was only exposed by changes in
to `transTerminatorKind` could lead to `TyFnDef`s shaped like variables falling
through to the fallback `Call` case, which was not equipped to handle `TyFnDef`s
- it expected `TyFnPtr`s. We resolve this by making `transTerminatorKind` check
callees by type, rather than by shape.