crucible-llvm: Support function pointer alignment in data layouts
This bumps the `llvm-pretty` submodule to bring in changes necessary to support
function pointer alignment in data layouts. It also adjusts the data
layout-parsing code in `crucible-llvm` accordingly. For now, the only
meaningful use of function pointer alignment in `crucible-llvm` is in computing
the `maxAlignment`.

See https://github.com/GaloisInc/llvm-pretty-bc-parser/issues/292 for the
parent issue involving function pointer alignment in data layouts.