crucible-mir: remove function pointer shim `CustomOp`s
At some point, these were necessary because `rustc` wouldn't generate MIR
implementations for `FnPtrShim`s on its own. I'm not sure when this changed in
`rustc`, though I would guess that it changed for us when we switched to Rust
1.86.

GHC did not alert me that `fnPtrShimDef` was no longer used, so I also added in
a `-Wunused-top-binds` to avoid that in the future.