crucible: Make `CrucibleType` into `type data`
`-XTypeData` enables the declaration of type-level datatypes without
the need for `-XDataKinds`-style "promotion" from the value level to
the type level. `CrucibleType` is only ever used at the type level, so
it is a prime candidate for a declaration as `type data`. In fact, it
was one of the use-cases for the GHC proposal introducing the extension!
Practically speaking, this lets us reduce a bunch of duplication and
boilerplate.

This functionality was introduced in GHC 9.6, so using it explicitly
drops support for GHC 9.4. This is in line with our versioning policy
as described in `doc/dev.md`, which states that we aim to support the
three most recent versions of GHC at a time. We currently support GHC
9.6-9.10.

https://github.com/ghc-proposals/ghc-proposals/blob/3b4e702f217884cf1c2d9cfbd9cb22ad642b22b0/proposals/0106-type-data.rst
https://downloads.haskell.org/ghc/9.6.1/docs/users_guide/9.6.1-notes.html