ci: Check freeze files on Linux only
As noticed on https://github.com/GaloisInc/.github/pull/59, the freeze
files may be missing libraries that are required to build on Windows
(e.g., Win32). The proper solution would likely be to check in freeze
files for each different platform. This would ensure deterministic
builds on each platform. However, this would impose an undue maintenance
burden, as not all developers have access to dev machines with each
supported OS installed. So, only check the completeness of the freeze
files on Linux.