crux-mir: Correctly concretize nested references
Previously, the implementation of `crucible::concretize` was incorrect for
nested reference types (e.g., `&&u32`). This is because when concretizing the
reference type, we must also recursively concretize the pointee value, but we
failed to propagate global-state changes from the recursive call correctly.
This leads to the global state missing `RefCell`s that it should have, as seen
in the oddities in #1513.

This fixes the issue by ensuring that we read from the newly updated global
state properly immediately after the recursive call.

Fixes #1513.