crux-mir: Don't add failing assertions for unsatisfiable paths in crucible::concretize
Previously, if we encountered an unsatisfiable path condition when attempting
to get the concrete model in `crucible::concretize`, it would add a failing
Crucible assertion. This is not quite what we want, however, as this can wreak
havoc if the call to `crucible::concretize` is in the middle of a solver frame,
as seen in the oddities observed in #1526.

This patch tweaks the `crucible::concretize` implementation to simply return
the original argument unchanged if the path condition is unsatisfiable. See the
newly added comments for a justification of why this is sensible.

Fixes #1526.