crucible-mir: Properly support const generics
This contains the `crucible-mir`-side changes needed to properly support const
generics, as detailed in https://github.com/GaloisInc/mir-json/issues/64.
Specifically, the `TyConst` data constructor now supports a `ConstVal` field
that describes the constant used to instantiate a const generic parameter. This
has no impact on the semantics of `crucible-mir` itself, but this `ConstVal`
can be used by downstream tools that wish to distinguish different
instantiations of const generic functions. (The primary use case is SAW's
`mir_find_adt` function.)

This bumps the `mir-json` submodule to bring in the changes from
https://github.com/GaloisInc/mir-json/pull/173. Because this requires a bump
to the MIR JSON schema version, corresponding changes to the version number
must also be made here.