Don't check -Wunticked-promoted-constructors
This warning is considered something of a misfeature. It was removed from
`-Wall` in GHC 9.4, and the opposite style (omitting ticks on promoted data
constructors unless it is absolutely necessary) is now encouraged. See
https://gitlab.haskell.org/ghc/ghc/-/issues/20531.

This removes all uses of `-Werror=unticked-promoted-constructors` from the
`.cabal` files in the repo. Moreover, it also removes some particularly
problematic uses of promoted data constructor ticks in
`crucible-mir:Mir.Intrinsics` that incorrectly cause `-Woperator-whitespace`
warnings to be emitted due to a GHC bug
(https://gitlab.haskell.org/ghc/ghc/-/issues/23297).