crucible-llvm: Fix the semantics of the llvm.is.fpclass override
Fix a couple of subtle typos (introduced in
https://github.com/GaloisInc/crucible/commit/6eca1935b6e7d1b56934ec3d03d2c9f5091f2760)
that causes the definitions of 1-bit zero and one to be swapped, resulting in
the semantics for the `llvm.is.fpclass` intrinsic to be utterly wrong. This
allows the `llvm.is.fpclass` test case to succeed on Clang 17 or later:

```
$ PATH=~/Software/clang+llvm-17.0.2/bin:$PATH cabal -v0 run crux-llvm-test -- -p llvm.is.fpclass
<snip>
crux-llvm
  clang 17.0.2
    llvm.is.fpclass.ll
      solver=z3
        loop-merging=loop
          clang-range=recent-clang
            z3 4.8.14
              llvm.is.fpclass
                llvm.is.fpclass crux run:    OK (0.14s)
                llvm.is.fpclass crux output: OK
        loop-merging=loopmerge
          clang-range=recent-clang
            z3 4.8.14
              llvm.is.fpclass
                llvm.is.fpclass crux run:    OK (0.15s)
                llvm.is.fpclass crux output: OK

All 4 tests passed (0.29s)
```

Fixes #1586.