crucible-mir: Improve error messages involving MIR JSON schema version mismatches
Previously, `parseMIR` would parse the entirety of a MIR JSON file upfront, and
if there was an error during parsing, it would report that error before
attempting to check if the MIR JSON schema versions match. Most of the time,
this leads to inscrutable JSON parse errors instead of more comprehensible
schema version mismatch errors, which isn't an ideal user experience.

This commit tweaks `parseMIR` such that if it fails to parse the MIR JSON file,
it will check for mismatched schema versions _before_ it throws a JSON parse
error.

Fixes #1618.