crux-llvm-test: Make tests pass with LLVM 19
There is one test case that did not pass out of the box with Clang 19, namely
`double_free.c`. This is for a very silly reason: when compiled with Clang 19,
the double-free error message mentions a pointer with block number `9`, but
with earlier versions of Clang, the error message mentions a pointer with block
number `10`. This is simply because earlier Clangs declare an `llvm.dbg.*`
intrinsic function that Clang 19 does not, and because each `declare`d function
in an LLVM file allocates a unique block number for the corresponding function
pointer, the lack of an additional `declare`d function with Clang 19 explains
why the error message has a block number decremented by one.