crux-mir-test: Ensure that calls to clone() aren't optimized away
`rustc` is known to optimize away bare calls to `clone()` in certain cases,
which would prevent certain `crux-mir` test cases from actually testing the
`clone()`-related code that they are meant to test. To make it less likely for
this to occur, guard the calls to `clone()` behind a wrapper function. Also
mark it as `#[inline(never)]` just to be extra cautious.