crucible-mir: Limited support for translating coroutines
Although we stop short of full support for `async`-related coroutine code (see
issue #1369 for that task), `crucible-mir` now has the ability to translate
coroutine closure expression and a limited ability to translate coroutine
expressions. This means that we can translate code involving coroutines without
`crucible-mir` crashing outright, instead giving a nicer `Coroutines not
currently supported` error message in the event that a user tries to simulate a
coroutine expression.

This bumps the `mir-json` submodule to bring in the corresponding changes from
https://github.com/GaloisInc/mir-json/pull/205. As a consequence, this requires
bumping the supported MIR JSON schema version.

Fixes #1626.