crucible-mir: monomorphize "subindex" indexing operations
This removes `subindexRef`, `MirSubindexRef`, et al. and replaces them with
separate operations that apply to specific representations. In particular, it
creates `mirRef_{vec,arr}Elem` for Crucible `Vector`s and `Array` access,
respectively, and uses the already-implemented `mirRef_agElem` for
`MirAggregate` access.

Why not keep `subindex` around and have it multiply the index by the type width
in aggregate access? Having the caller provide the type width is awkward, since
it isn't needed in the vector or array case. It's also awkward to store a `Ty`
in the `SubindexRef` node to calculate the width from it, since converting it to
a `TypeRepr` is necessary eventually, and doing it in the most natural place
(`subindexMirRefLeaf`) requires a `tyToRepr` import that results in a module
import cycle.

NFCI